got rid of robot centric drive and almost done w auto
